full_automation/requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
secure-smtplib
